+++ ServiceWebsite/ServiceWebsite/ClientApp/src/app/modules/individual-journey/pages/about-you/about-you.component.spec.ts
@@ -1,6 +1,6 @@
 import { IndividualJourneyComponentTestBed } from '../individual-base-component/individual-component-test-bed.spec';
 import { AboutYouComponent } from './about-you.component';
-import { SuitabilityChoiceComponentFixture } from 'src/app/modules/base-journey/components/suitability-choice-page-base.component.spec';
+import { SuitabilityChoiceComponentFixture } from 'src/app/modules/base-journey/components/suitability-choice-component-fixture.spec';
 
 describe('AboutYouComponent', () => {
   let fixture: SuitabilityChoiceComponentFixture;
