+++ ServiceWebsite/ServiceWebsite/ClientApp/src/app/modules/representative-journey/pages/about-you/about-you.component.spec.ts
@@ -2,7 +2,7 @@ import { AboutYouComponent } from './about-you.component';
 
 import {
   SuitabilityChoiceComponentFixture
-} from 'src/app/modules/base-journey/components/suitability-choice-page-base.component.spec';
+} from 'src/app/modules/base-journey/components/suitability-choice-component-fixture.spec';
 import { RepresentativeJourneyComponentTestBed } from '../representative-base-component/component-test-bed.spec';
 
 describe('AboutYouComponent', () => {
